Add files via upload
Creating a classification algorithm for audio data as the final project for the CS50 Computer Science course from Harvard.
Participants hummed songs as training data for various ML algorithms to train on. In the end, the most accurate algorithm was chosen to be uploaded to a website to identify further hummings of the same songs.